# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WhaleWatch API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/filings/{accession_no}": {
      "get": {
        "tags": [
          "filings"
        ],
        "summary": "One filing, with its holdings",
        "description": "Read a filing and every position it reports, largest first.\n\nTwo queries rather than one: a filing joined to its holdings repeats every\nfiling column on every row, and a large 13F is three thousand rows. The\nsecond query is a lookup on ``ix_holding_filing_id``.",
        "operationId": "getFiling",
        "parameters": [
          {
            "name": "accession_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "EDGAR accession number, with or without dashes: `##########-##-######`. Both spellings are in circulation — dashed in EDGAR's indexes, undashed in archive URLs — and both reach the same filing.",
              "examples": [
                "0001067983-24-000011"
              ],
              "title": "Accession No"
            },
            "description": "EDGAR accession number, with or without dashes: `##########-##-######`. Both spellings are in circulation — dashed in EDGAR's indexes, undashed in archive URLs — and both reach the same filing."
          },
          {
            "name": "include_options",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Set `false` to drop option lines (`put_call` is not null). Their `value_usd` is the notional value of the underlying rather than a premium, so a portfolio total that includes them is inflated by the whole exposure.",
              "default": true,
              "title": "Include Options"
            },
            "description": "Set `false` to drop option lines (`put_call` is not null). Their `value_usd` is the notional value of the underlying rather than a premium, so a portfolio total that includes them is inflated by the whole exposure."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilingRead"
                }
              }
            }
          },
          "404": {
            "description": "No filing with that accession number.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "A parameter is missing or malformed. `errors` says which.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AmendmentKind": {
        "type": "string",
        "enum": [
          "restatement",
          "new_holdings"
        ],
        "title": "AmendmentKind",
        "description": "What a ``13F-HR/A`` claims to do to the period it amends.\n\nThis is the single most consequential field on an amendment, and getting it\nbackwards is silent. A **restatement** replaces the period's holdings\nwholesale: load it alongside the original and every position is counted\ntwice. A **new holdings** amendment adds rows the original omitted —\ntypically positions released from confidential treatment — and treating\n*that* as a restatement throws away everything the original reported.\n\nBoth failures produce a portfolio that looks entirely plausible. Neither\nraises anything.\n\nThe values are the normalised form of EDGAR's own ``<amendmentType>``:\n``RESTATEMENT`` and ``NEW HOLDINGS``. Normalised rather than raw because the\nspace in the second one makes every SQL literal in the codebase quotable but\nugly, and because these two are a closed set — unlike ``transaction_code``\non Form 4, which is stored exactly as filed precisely because it is not.\n\n``None`` on the column is the third state and the common one: the filing is\nnot an amendment at all."
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "invalid_cursor",
          "not_found",
          "method_not_allowed",
          "validation_error",
          "http_error"
        ],
        "title": "ErrorCode",
        "description": "What went wrong, for a program to branch on. ``detail`` says it for a person."
      },
...
      "FieldError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Loc",
            "description": "Where the parameter is, then its name: `[\"query\", \"period\"]`.",
            "examples": [
              [
                "query",
                "period"
              ]
            ]
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "description": "Why it was refused, for a person.",
            "examples": [
              "Value error, '2026Q5' has no quarter 5: quarters are 1 to 4"
            ]
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Pydantic's error type, such as `missing` or `value_error`.",
            "examples": [
              "value_error"
            ]
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "FieldError",
        "description": "One parameter that failed validation, and why."
      },
...
      "FilingRead": {
        "properties": {
          "accession_no": {
            "type": "string",
            "title": "Accession No",
            "description": "EDGAR's identifier, dashed: `##########-##-######`.",
            "examples": [
              "0001067983-24-000011"
            ]
          },
          "cik": {
            "type": "string",
            "title": "Cik",
            "description": "As EDGAR filed it, zero-padded to ten. Not necessarily a resolved filer.",
            "examples": [
              "0001067983"
            ]
          },
          "form_type": {
            "type": "string",
            "title": "Form Type",
            "description": "`13F-HR`, `13F-HR/A` for an amendment, or `13F-NT` for a notice.",
            "examples": [
              "13F-HR"
            ]
          },
          "period_of_report": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Of Report",
            "description": "The quarter end being reported — never the filing date, which is 45 days or more later.",
            "examples": [
              "2024-03-31"
            ]
          },
          "quarter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quarter",
            "description": "Generated by Postgres from `period_of_report`; null exactly when it is.",
            "examples": [
              "2024Q1"
            ]
          },
          "filed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Filed At",
            "description": "When it reached EDGAR.",
            "examples": [
              "2024-05-15T16:05:22Z"
            ]
          },
          "filer_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filer Id",
            "description": "Null until this filing's CIK has been resolved to a filer.",
            "examples": [
              1
            ]
          },
          "filer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filer Name",
            "description": "The resolved filer's name. Null where `filer_id` is.",
            "examples": [
              "Berkshire Hathaway"
            ]
          },
          "filer_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filer Slug",
            "description": "The resolved filer's slug: `/v1/investors/{slug}`. Null where `filer_id` is.",
            "examples": [
              "berkshire-hathaway"
            ]
          },
          "value_multiplier": {
            "type": "integer",
            "title": "Value Multiplier",
            "description": "Which units the filing's own `value` field used: 1000 before the 2023-01-03 cutover, 1 on and after it. Recorded, not applied — `holdings[].value_usd` is already in whole dollars either way. This is the field to check first when a fund's total is out by 1000x.",
            "examples": [
              1
            ]
          },
          "amendment_no": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amendment No",
            "description": "The cover page's amendment number, as the filer wrote it. Null on an original. Amendments apply in `filed_at` order, not in this order.",
            "examples": [
              1
            ]
          },
          "amendment_kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AmendmentKind"
              },
              {
                "type": "null"
              }
            ],
            "description": "`restatement` replaces the original period; `new_holdings` adds to it.",
            "examples": [
              "restatement"
            ]
          },
          "amends_accession_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amends Accession No",
            "description": "The filing this one amends, if any.",
            "examples": [
              "0000950123-24-005032"
            ]
          },
          "report_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Type",
            "description": "Cover-page wording. A `NOTICE` reports no holdings at all; a `COMBINATION` reports only the part this manager holds directly.",
            "examples": [
              "13F HOLDINGS REPORT"
            ]
          },
          "parse_status": {
            "type": "string",
            "title": "Parse Status",
            "description": "`pending` (fetched, not parsed), `ok`, `suspect` (loaded, but a normalisation guard fired — see `parse_notes`), or `failed` (not parsed at all; `parse_error` says why and there are no holdings).",
            "examples": [
              "ok"
            ]
          },
          "parse_notes": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parse Notes",
            "description": "Why a filing is `suspect`: one object per finding, with the row and CUSIP. `severity` is `error` for a guard that failed, which is what makes the filing `suspect`, or `warning` for evidence beside it.",
            "examples": [
              [
                {
                  "cusip": "037833100",
                  "detail": "...",
                  "kind": "implied_price",
                  "severity": "error"
                }
              ]
            ]
          },
          "parse_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parse Error",
            "description": "Set only alongside `parse_status: failed`.",
            "examples": [
              null
            ]
          },
          "parsed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parsed At",
            "description": "When it was last parsed. Null while `parse_status` is `pending`.",
            "examples": [
              "2024-05-15T16:20:41Z"
            ]
          },
          "ingested_at": {
            "type": "string",
            "format": "date-time",
            "title": "Ingested At",
            "description": "When we first fetched it.",
            "examples": [
              "2024-05-15T16:20:39Z"
            ]
          },
          "raw_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Key",
            "description": "Where the archived documents live: a prefix in the raw store.",
            "examples": [
              "raw/13f/1067983/0001067983-24-000011/"
            ]
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url",
            "description": "The URL it was fetched from, stored rather than reconstructed.",
            "examples": [
              "https://www.sec.gov/Archives/edgar/data/1067983/000106798324000011/"
            ]
          },
          "holdings": {
            "items": {
              "$ref": "#/components/schemas/HoldingRead"
            },
            "type": "array",
            "title": "Holdings",
            "description": "Sorted by `value_usd` descending, largest position first. Empty for a 13F-NT, for a filing that failed to parse, and for one whose holdings have not been loaded yet."
          }
        },
        "type": "object",
        "required": [
          "accession_no",
          "cik",
          "form_type",
          "filed_at",
          "value_multiplier",
          "parse_status",
          "ingested_at"
        ],
        "title": "FilingRead",
        "description": "A filing, its provenance, and every position it reports."
      },
...
      "HoldingRead": {
        "properties": {
          "cusip": {
            "type": "string",
            "title": "Cusip",
            "description": "The nine characters this filing reported, as it reported them.",
            "examples": [
              "037833100"
            ]
          },
          "issuer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issuer Name",
            "description": "From `security`, our resolution — not the issuer name this filing printed.",
            "examples": [
              "APPLE INC"
            ]
          },
          "ticker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ticker",
            "description": "From `security`, our resolution. Null while the CUSIP is unresolved.",
            "examples": [
              "AAPL"
            ]
          },
          "value_usd": {
            "type": "string",
            "format": "decimal",
            "title": "Value Usd",
            "description": "Whole dollars, always — already multiplied through by `value_multiplier`. Not comparable across `put_call`: an option line reports the notional value of the underlying, not a premium.",
            "examples": [
              "2040000000.00"
            ]
          },
          "shares": {
            "type": "string",
            "format": "decimal",
            "title": "Shares",
            "description": "A share count or a principal amount; `sshprnamt_type` says which.",
            "examples": [
              "12000000.0000"
            ]
          },
          "sshprnamt_type": {
            "type": "string",
            "title": "Sshprnamt Type",
            "description": "`SH` for shares, `PRN` for principal.",
            "examples": [
              "SH"
            ]
          },
          "put_call": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Put Call",
            "description": "`Put`, `Call`, or null for the underlying itself.",
            "examples": [
              null
            ]
          },
          "investment_discretion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Investment Discretion",
            "description": "As the filing wrote it: `SOLE`, `DFND` (shared, defined) or `OTR` (shared, other).",
            "examples": [
              "SOLE"
            ]
          },
          "voting_sole": {
            "anyOf": [
              {
                "type": "string",
                "format": "decimal"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voting Sole",
            "description": "Shares the manager votes alone. Null when the filing left it out.",
            "examples": [
              "12000000.0000"
            ]
          },
          "voting_shared": {
            "anyOf": [
              {
                "type": "string",
                "format": "decimal"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voting Shared",
            "description": "Shares the manager votes with others. Null when the filing left it out.",
            "examples": [
              "0.0000"
            ]
          },
          "voting_none": {
            "anyOf": [
              {
                "type": "string",
                "format": "decimal"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voting None",
            "description": "Shares the manager does not vote. Null when the filing left it out.",
            "examples": [
              "0.0000"
            ]
          }
        },
        "type": "object",
        "required": [
          "cusip",
          "value_usd",
          "shares",
          "sshprnamt_type"
        ],
        "title": "HoldingRead",
        "description": "One reported position, as the filing wrote it plus what we resolved it to.\n\n``cusip`` is the filing's own nine characters and ``issuer_name``/``ticker``\nare our interpretation of them, which is exactly the pair you want side by\nside when a position is attributed to the wrong company: a null ticker is an\nunresolved security, a wrong one is a bad resolution, and the CUSIP tells\nyou which of the two you are looking at."
      },
...
      "Problem": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "What went wrong, stable across releases. `invalid_cursor`: `cursor` is malformed, or belongs to a different listing or sort. `not_found`: no such thing, or nothing published for the period asked for. `validation_error`: a parameter is missing or malformed; `errors` says which.",
            "examples": [
              "not_found"
            ]
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "What went wrong, in a sentence for a person. Not a contract: do not parse it.",
            "examples": [
              "No investor 'berkshire'. GET /v1/investors lists every one."
            ]
          }
        },
        "type": "object",
        "required": [
          "code",
          "detail"
        ],
        "title": "Problem",
        "description": "Why the request was refused."
      },
...
      "ValidationProblem": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "What went wrong, stable across releases. `invalid_cursor`: `cursor` is malformed, or belongs to a different listing or sort. `not_found`: no such thing, or nothing published for the period asked for. `validation_error`: a parameter is missing or malformed; `errors` says which.",
            "examples": [
              "not_found"
            ]
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "What went wrong, in a sentence for a person. Not a contract: do not parse it.",
            "examples": [
              "No investor 'berkshire'. GET /v1/investors lists every one."
            ]
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array",
            "title": "Errors",
            "description": "Every parameter that failed, each once. Never empty."
          }
        },
        "type": "object",
        "required": [
          "code",
          "detail",
          "errors"
        ],
        "title": "ValidationProblem",
        "description": "The 422 body: a parameter is missing or malformed."
      }
    }
  }
}